fix: move pure orchestrate handlers into MCP server rootDir
Pure TypeScript gate handlers in src/orchestrate/ were imported by
the MCP server via ../../../../src/orchestrate/ paths, violating
TypeScript's rootDir constraint (TS6059) in CI.

Relocate to servers/exarchos-mcp/src/orchestrate/pure/ and update
all imports and vi.mock() paths. No logic changes.
