+++ servers/exarchos-mcp/src/orchestrate/pure/context-economy.test.ts
