+++ servers/exarchos-mcp/src/orchestrate/pure/context-economy.ts
