+++ servers/exarchos-mcp/src/orchestrate/pure/design-completeness.test.ts
