+++ servers/exarchos-mcp/src/orchestrate/pure/operational-resilience.ts
