+++ servers/exarchos-mcp/src/orchestrate/pure/provenance-chain.ts
