+++ servers/exarchos-mcp/src/orchestrate/pure/static-analysis.test.ts
