+++ servers/exarchos-mcp/src/orchestrate/pure/static-analysis.ts
