+++ servers/exarchos-mcp/src/orchestrate/static-analysis.ts
@@ -9,8 +9,8 @@ import { execFileSync } from 'node:child_process';
 import type { ToolResult } from '../format.js';
 import { getOrCreateEventStore } from '../views/tools.js';
 import { emitGateEvent } from './gate-utils.js';
-import { runStaticAnalysis } from '../../../../src/orchestrate/static-analysis.js';
-import type { RunCommandFn, CommandResult } from '../../../../src/orchestrate/static-analysis.js';
+import { runStaticAnalysis } from './pure/static-analysis.js';
+import type { RunCommandFn, CommandResult } from './pure/static-analysis.js';
 
 // ─── Argument & Result Types ─────────────────────────────────────────────────
